backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
celery[redis]==5.4.0
redis==5.2.0
anthropic==0.42.0
cloudinary==1.41.0
pgvector==0.3.6
python-multipart==0.0.18
httpx==0.28.0
pytest==8.3.0
pytest-asyncio==0.24.0
resend==2.0.0
numpy>=1.24.0,<2.1.0
